Modelação de dados do mercado
